Fixed bug with inherited VIRTUAL_ENV variable
When shell was run with su from parent shell that was in a direcory with
virtualenv, new shell inherited VIRTUAL_ENV variable from parent shell
and script thought virtualenv is activated, when in new shell it wasn't.

That could have been fixed with improved check, unsetting VIRTUAL_ENV
seems to be right.